feat: Add missing functionalities
Added the 'courses' variable to the service interface, which returns an
array of course objects. This variable access to the missing
functionalities in the course service.

Implemented the ''getRoundedPrice' method, which take price and
installment count as parameters and returns the rounded result after the
dividing the price by the number as installments.